# feat: defer libuv calls in Path:new