Reduces `Path:new` to c. 1/3 of cost. Can be quite relevant for `telescope` finders that more greedily create `Path`s. Just thought I'd put this up for discussion.

```lua
local bench = require "plenary.benchmark"
local Path = require "plenary.path"

-- /home/fdschmidt/repos/lua/telescope-file-browser.nvim/
local filename = vim.api.nvim_buf_get_name(0)

bench("Path", {
  warmup = 10,
  runs = 50,
  fun = {
    {
      "Path creation",
      function()
        for _ = 1, 1000 do
          Path:new(filename)
        end
      end,
    },
  },
})
```

Before

Benchmark Group: 'Path' -----------------------                                                                                                                                                 
Benchmark 1: 'Path creation'                                                                                                                                                                   
  Time(mean ± σ):      9.2 ms ±   2.9 ms                                                                                                                                                        
  Range(min … max):    7.6 ms …  17.2 ms  50 runs

After

Benchmark Group: 'Path' -----------------------                                                                                                                                              
Benchmark 1: 'Path creation'                                                                                                                                                                
  Time(mean ± σ):      3.6 ms ±   1.0 ms                                                                                                                                                     
  Range(min … max):    3.0 ms …   7.8 ms  50 runs  

cc @miversen33