@miversen33 2022-05-21
These changes will _actively_ make my life easier with Netman! You're the best :)